task-scheduler/requirements.txt
numpy==1.24.3
pandas==2.0.3
python-redmine==2.4.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0 